Added - Basic header/source files for player
This will probably be overwritten and modified later. This is just a
base template for implementing the player character somehow.
- player.h extends Tile class. So this means that the player instance
would be a tile in the tilemap...

for player animations it may be problematic if the player is only a tile
instance... but it could work depending on how you update each character
piece.